% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_analyte_polluter.R
\name{add_new_analyte_polluter}
\alias{add_new_analyte_polluter}
\title{Add new water chemistry data and polluting events at new locations}
\usage{
add_new_analyte_polluter(
  path_total,
  df_analyte_to_append_filepath,
  df_polluter_to_append_file_path,
  upstream_thresh = 5,
  downstream_lower_thresh = 0,
  downstream_upper_thresh = 50,
  num_cores = 1,
  permission = "no"
)
}
\arguments{
\item{path_total}{This is the same root file path as that of the function “whole_pipe”. The previous output will be overwritten so backup is recommended.}

\item{df_analyte_to_append_filepath}{This is the full path to the csv file containing new water chemistry data for the given stream. The file formatting requirement is the same as that of “filename_chem” in “whole_pipe”.}

\item{df_polluter_to_append_file_path}{This is the full path to the csv file containing new pollution site data for the given stream. The file formatting requirement is the same as that of “filename_poll” in “whole_pipe”.}

\item{upstream_thresh}{The same parameter as that in "whole_pipe".}

\item{downstream_lower_thresh}{The same parameter as that in "whole_pipe".}

\item{downstream_upper_thresh}{The same parameter as that in "whole_pipe".}

\item{num_cores}{The same parameter as that in "whole_pipe".}

\item{permission}{The same parameter as that in "whole_pipe".}
}
\value{
The same return files as those in "whole_pipe".
}
\description{
This function can be used to add new water chemistry data and polluting events at new locations and run the statistical inference for all the polluting events. This function can only be used after “whole_pipe” is executed. Execution of this function will overwrite the previously generated output. Therefore, it is advised to preserve a copy of the old outputs in case you need it.
}
\note{
This function assumes “whole_pipe” function has already been run and the outputs generated from it lie in “path_total” folder. Please refer to our \href{https://github.com/HANDS-Research-Group/StreamNet/tree/main/Rpackage_StreamNet}{website} for examples.
}
