% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_link_length.R
\name{avg_link_length}
\alias{avg_link_length}
\title{Calculate average link length of stream network}
\usage{
avg_link_length(lines, outlet_reach)
}
\arguments{
\item{lines}{sf linestring collection}
}
\description{
Calculate average link length of stream network
}
