% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_distance.R
\name{get_focal_distance}
\alias{get_focal_distance}
\title{Get the network distance between a focal point and other points on a line network}
\usage{
get_focal_distance(lines, lakes, outlet, size_threshold = 4)
}
\arguments{
\item{lines}{sf lines object}

\item{lakes}{sf polygon object}

\item{outlet}{integer row index of outlet reach relative to lines}

\item{size_threshold}{numeric size above which to consider as a lake}
}
\description{
Get the network distance between a focal point and other points on a line network
}
\examples{
\dontrun{
library(nhdR)

data(nhd_sub_lines)
data(nhd_sub_lakes)

outlet_reach   <- terminal_reaches(network = nhd_sub_lines,
                                   approve_all_dl = TRUE)
outlet <- which(outlet_reach[['comid']] == nhd_sub_lines[['comid']])

get_focal_distance(nhd_sub_lines, nhd_sub_lakes, outlet = outlet)

}

}
