% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake_distance.R
\name{get_mat_distance}
\alias{get_mat_distance}
\title{Get the network distance matrix between points on a line network}
\usage{
get_mat_distance(lines, lakes, size_threshold = 4)
}
\arguments{
\item{lines}{sf lines object}

\item{lakes}{sf polygon object}

\item{size_threshold}{numeric size above which to consider as a lake}
}
\description{
Get the network distance matrix between points on a line network
}
\examples{
\dontrun{
library(nhdR)

data(nhd_sub_lines)
data(nhd_sub_lakes)

get_mat_distance(nhd_sub_lines, nhd_sub_lakes)
}

}
