% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrolinks.R
\name{hlk_traverse}
\alias{hlk_traverse}
\title{Retrieve stream network draining to a given lake}
\usage{
hlk_traverse(wb_coords, dset = "nhdh")
}
\arguments{
\item{wb_coords}{numeric vector of length 2: lat, lon}

\item{dset}{character dataset identifier: nhdh, nhdplusv2}
}
\description{
Retrieve stream network draining to a given lake
}
\examples{
\dontrun{
wb_coords <- c(44.00467, -88.43445)
res <- hlk_traverse(wb_coords)
}
}
