% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{igraph2sf}
\alias{igraph2sf}
\title{Convert an igraph object to sf}
\usage{
igraph2sf(ig_network)
}
\arguments{
\item{ig_network}{igraph network}
}
\description{
Convert an igraph object to sf
}
\examples{
library(sf)
tree <- create_reversed_tree(15)
tree_sf <- igraph2sf(tree)
plot(tree_sf)
}
