% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spnetwork.R
\name{sf2network}
\alias{sf2network}
\title{Convert sf to SpatialNetwork}
\usage{
sf2network(sf_lines)
}
\arguments{
\item{sf_lines}{sf lines object}
}
\description{
Convert sf to SpatialNetwork
}
\examples{
\dontrun{
# library(igraph)
# library(sp)
# library(mapview)
data(nhd_sub)
res <- sf2network(nhd_sub)
plot(res@g)
}
}
