% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_link_length.R
\name{simplify_network}
\alias{simplify_network}
\title{Simplify network}
\usage{
simplify_network(lines)
}
\arguments{
\item{lines}{sf data.frame composed of LINESTRING objects}
}
\description{
Combine(dissolve) adjacent reaches with no junctions
}
\examples{
\dontrun{
data(nhd_sub_lines)
res <- simplify_network(nhd_sub_lines)

# single line returns the original
res <- simplify_network(nhd_sub_lines[1,])

}
}
