% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_order.R
\name{stream_order}
\alias{stream_order}
\title{stream_order}
\usage{
stream_order(lines, outlet, ...)
}
\arguments{
\item{lines}{lines}

\item{outlet}{outlet}

\item{...}{options passed to rgrass7sf::initGRASS}
}
\description{
Calculate stream order
}
