% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_order.R
\name{stream_order_ratio}
\alias{stream_order_ratio}
\title{Calculate stream order ratio with GRASS}
\usage{
stream_order_ratio(lines, outlet, ...)
}
\arguments{
\item{lines}{lines}

\item{outlet}{outlet}

\item{...}{options passed to rgrass7sf::initGRASS}
}
\description{
Calculate stream order ratio with GRASS
}
\examples{
\dontrun{
library(sf)

data(nhd_sub)

outlet_reach   <- terminal_reaches(network = nhd_sub, approve_all_dl = TRUE)
outlet_point   <- st_cast(st_line_sample(outlet_reach, sample = 1), "POINT")

stream_order_ratio(lines = nhd_sub, outlet = outlet_point)
}
}
