% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{viz_dla}
\alias{viz_dla}
\title{Vizualize the dla simulation}
\usage{
viz_dla(r, origin)
}
\arguments{
\item{r}{raster made from output of sim_dla}

\item{origin}{integer index of network origin}
}
\description{
Vizualize the dla simulation
}
\examples{
\dontrun{
dt <- sim_dla()
viz_dla(raster(dt), which.max(dt))
image(dt)
}
}
