% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverdist.R
\name{autoclean}
\alias{autoclean}
\title{Automatically clean a rivernetwork object}
\usage{
autoclean(rivernetwork, mouthseg, mouthvert, crs)
}
\arguments{
\item{rivernetwork}{rivernetwork object from the riverdist package}

\item{mouthseg}{integer}

\item{mouthvert}{integer}

\item{crs}{crs string or epsg code}
}
\description{
Automatically clean a rivernetwork object
}
\examples{
\dontrun{
mouthseg  <- which(nhd$COMID == outlet_reach$comid)
res <- autoclean(nhd_rv, mouthseg, mouthvert, crs = st_crs(nhd))
}
}
