% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_link_length.R
\name{avg_link_length}
\alias{avg_link_length}
\title{Calculate average link length of stream network}
\usage{
avg_link_length(lines, simplify = FALSE)
}
\arguments{
\item{lines}{sf linestring collection}

\item{simplify}{logical run simplify_network prior to calculations?}
}
\description{
Calculate average link length of stream network
}
\examples{
\dontrun{
data(nhd_sub)
avg_link_length(nhd_sub)
}
}
