% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamnet-package.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Calculate connectivity metrics}
\usage{
calc_metrics(lines, lakes, map = FALSE)
}
\arguments{
\item{lines}{sf lines object}

\item{lakes}{sf polygon object}

\item{map}{logical show a map output of the results?}
}
\description{
Calculate connectivity metrics
}
\examples{
\dontrun{

data(nhd_sub_lines)
data(nhd_sub_lakes)

res <- calc_metrics(nhd_sub_lines, nhd_sub_lakes)

}
}
