% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_reversed_tree}
\alias{create_reversed_tree}
\title{Create a reversed tree igraph object}
\usage{
create_reversed_tree(node_n)
}
\arguments{
\item{node_n}{number of nodes}
}
\description{
Create a reversed tree igraph object
}
\examples{
create_reversed_tree(5)
}
