% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverdist.R
\name{rvnet2sf}
\alias{rvnet2sf}
\title{Rivernetwork to sf object}
\usage{
rvnet2sf(rvnet)
}
\arguments{
\item{rvnet}{rivernetwork object from the riverdist package}
}
\description{
Rivernetwork to sf object
}
