% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{sf2igraph}
\alias{sf2igraph}
\title{Convert sf lines to an igraph object}
\usage{
sf2igraph(sf_lines, tolerance = 1)
}
\arguments{
\item{sf_lines}{}
}
\description{
Convert sf lines to an igraph object
}
\examples{
tree    <- create_reversed_tree(15)
tree_sf <- igraph2sf(tree)
tree    <- sf2igraph(tree_sf, tolerance = 1)
plot(tree$tree,
     edge.width = tree$weights,
     layout = igraph::layout_as_tree(tree$tree, mode = "in"))
}
