% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spnetwork.R
\name{sf2network}
\alias{sf2network}
\title{Convert sf lines to SpatialNetwork}
\usage{
sf2network(sf_lines)
}
\arguments{
\item{sf_lines}{sf lines object}
}
\description{
Convert sf lines to SpatialNetwork
}
\examples{
\dontrun{
data(nhd_sub_lines)
res <- sf2network(nhd_sub_lines)
plot(res)
}
}
