% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_order.R
\name{stream_order}
\alias{stream_order}
\title{Calculate stream order with GRASS}
\usage{
stream_order(lines, outlet, ...)
}
\arguments{
\item{lines}{lines}

\item{outlet}{outlet}

\item{...}{options passed to rgrass7sf::initGRASS}
}
\description{
Calculate stream order with GRASS
}
\examples{
\dontrun{
library(sf)
library(mapview)

data(nhd_sub)

outlet <- st_cast(st_line_sample(
              dplyr::filter(nhd_sub, comid == "7718290"),
          sample = 1), "POINT")

res <- stream_order(lines = nhd_sub, outlet = outlet)
mapview(res, zcol = "strahler")
}
}
